requirements.txt
redis>=5.0.0
python-dotenv>=1.0.0
bcrypt>=4.1.2
